Fix grow sitemap generation and skip sitemap for grow tests
